E3-DevMind-AI/requirements-agui.txt
# E3 DevMind AG-UI Server Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LangChain & LangGraph (all required components)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langgraph-checkpoint>=0.1.0
langgraph-prebuilt>=0.1.0

# AG-UI Protocol
ag-ui-protocol>=0.1.0
ag-ui-langgraph>=0.1.0
copilotkit>=0.1.0

# Async support
httpx>=0.27.0
aiohttp>=3.9.0

# Additional dependencies
openai>=1.0.0
tiktoken>=0.5.0
